Add --level_sections arg to run_solver.py
Level sections is the number of sections to split the generated level
into such that the start tile will be in the first section and the
goal tile will be in the last section. The default value is 2 if left
unspecified. This ensures an appropriate distance between the start and
goal tiles of generated levels.